implement connect in sdk
